requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
langchain-core
langchain-anthropic
langgraph
anthropic
